feat: refactor code and handle error when creating new user
