+++ src/user/user.module.ts
@@ -5,10 +5,13 @@ import { AuthModule } from 'src/auth/auth.module';
 import { AuthService } from 'src/auth/auth.service';
 import { ConfigModule } from '@nestjs/config';
 import { StorageModule } from 'src/storage/storage.module';
+import { UserAdminService } from './user.admin.service';
+import { UserFollowService } from './user.follow.service';
+import { UserAdminController } from './user.admin.controller';
 
 @Module({
   imports: [AuthModule, ConfigModule, StorageModule],
-  controllers: [UserController],
-  providers: [UserService, AuthService],
+  controllers: [UserController, UserAdminController],
+  providers: [UserService, UserAdminService, UserFollowService, AuthService],
 })
 export class UserModule {}
